# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Multi-Agent Fact Verification API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface InvestigateInput {
  /**
     * Text query to investigate
     * @nullable
     */
  query?: string | null;
  /**
     * Base64-encoded image for screenshot-based claim extraction
     * @nullable
     */
  image_base64?: string | null;
  /**
     * MIME type of the uploaded image (e.g. image/png, image/jpeg)
     * @nullable
     */
  image_mime_type?: string | null;
}

export interface SourceItem {
  url: string;
  title: string;
  /** 0-100 reliability score based on source type */
  reliability_score: number;
  /** government | academic | major_news | blog */
  source_type: string;
  /** @nullable */
  published_date?: string | null;
}

export interface VerifiedClaim {
  id: number;
  claim_text: string;
  /** Final confidence score 5-98 */
  confidence_score: number;
  /** Verified | Partially Verified | Low Verification */
  verdict: string;
  sources: SourceItem[];
  skeptic_note: string;
  low_verification: boolean;
  outdated_or_conflicting: boolean;
  avg_reliability: number;
  source_count: number;
}

export interface AgentStep {
  /** Human-readable agent name */
  agent: string;
  /** research | verification | skeptic | final */
  agent_id: string;
  message: string;
  timestamp: string;
}

export interface InvestigateResult {
  /** The resolved text query (extracted from image if applicable) */
  query: string;
  /**
     * Base64 thumbnail of uploaded image for display
     * @nullable
     */
  image_thumbnail?: string | null;
  claims: VerifiedClaim[];
  agent_trace: AgentStep[];
}

export interface ErrorResponse {
  error: string;
}