# lib/api-zod/src/generated/types/agentStep.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Multi-Agent Fact Verification API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentStep {
  /** Human-readable agent name */
  agent: string;
  /** research | verification | skeptic | final */
  agent_id: string;
  message: string;
  timestamp: string;
}