# lib/api-zod/src/generated/types/investigateInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Multi-Agent Fact Verification API
 * OpenAPI spec version: 0.1.0
 */

export interface InvestigateInput {
  /**
     * Text query to investigate
     * @nullable
     */
  query?: string | null;
  /**
     * Base64-encoded image for screenshot-based claim extraction
     * @nullable
     */
  image_base64?: string | null;
  /**
     * MIME type of the uploaded image (e.g. image/png, image/jpeg)
     * @nullable
     */
  image_mime_type?: string | null;
}