# lib/api-zod/src/generated/types/investigateResult.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Multi-Agent Fact Verification API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentStep } from './agentStep';
import type { VerifiedClaim } from './verifiedClaim';

export interface InvestigateResult {
  /** The resolved text query (extracted from image if applicable) */
  query: string;
  /**
     * Base64 thumbnail of uploaded image for display
     * @nullable
     */
  image_thumbnail?: string | null;
  claims: VerifiedClaim[];
  agent_trace: AgentStep[];
}