# lib/api-zod/src/generated/types/sourceItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Multi-Agent Fact Verification API
 * OpenAPI spec version: 0.1.0
 */

export interface SourceItem {
  url: string;
  title: string;
  /** 0-100 reliability score based on source type */
  reliability_score: number;
  /** government | academic | major_news | blog */
  source_type: string;
  /** @nullable */
  published_date?: string | null;
}