# lib/api-zod/src/generated/types/verifiedClaim.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Multi-Agent Fact Verification API
 * OpenAPI spec version: 0.1.0
 */
import type { SourceItem } from './sourceItem';

export interface VerifiedClaim {
  id: number;
  claim_text: string;
  /** Final confidence score 5-98 */
  confidence_score: number;
  /** Verified | Partially Verified | Low Verification */
  verdict: string;
  sources: SourceItem[];
  skeptic_note: string;
  low_verification: boolean;
  outdated_or_conflicting: boolean;
  avg_reliability: number;
  source_count: number;
}